requirements.txt
requests>=2.31.0
PyYAML>=6.0
pytz>=2023.3
colorama>=0.4.6
Flask>=3.0.0
Flask-CORS>=4.0.0
aiohttp>=3.9.0
tornado>=6.4
paramiko>=3.4.0
PyJWT>=2.8.0